+++ src/controllers/ScreenController.js
@@ -1,6 +1,7 @@
 import { db_close, db_connect } from "../database/db.js"
 import { cinemaModel } from "../schema/cinema.js"
 import { screenModel } from "../schema/screen.js"
+import { seatModel } from "../schema/seat.js"
 
 const getScreens = async (req, res) => {
     const cinema = req.query.cinema
@@ -8,10 +9,10 @@ const getScreens = async (req, res) => {
         await db_connect()
         let screens = []
         if (cinema) {
-            screens = await screenModel.find({ cinema: cinema }).exec()
+            screens = await screenModel.find({ cinema: cinema }).populate('seats').populate('showtimes').exec()
         }
         else {
-            screens = await screenModel.find({}).exec()
+            screens = await screenModel.find({}).populate('seats').populate('showtimes').exec()
         }
         res.status(200).json(screens)
     } catch (error) {
@@ -41,12 +42,18 @@ const createScreen = async (req, res) => {
     const { screenType, capacity, cinema } = req.body
     try {
         await db_connect()
-        const screen = await screenModel.create({ capacity, screenType, cinema })
-        await cinemaModel.findByIdAndUpdate(
-            cinema,
-            { $addToSet: { screens: screen._id } },
-        )
-        res.status(201).json(screen)
+        const cinema_exist = await cinemaModel.findById(cinema)
+        if (cinema_exist) {
+            const screen = await screenModel.create({ capacity, screenType, cinema })
+            await cinemaModel.findByIdAndUpdate(
+                cinema,
+                { $addToSet: { screens: screen._id } },
+            )
+            res.status(201).json(screen)
+        } else {
+            res.status(404).send("Cinema does not exist")
+        }
+
     } catch (error) {
         res.status(500).send(error.message)
     } finally {
@@ -77,8 +84,14 @@ const deleteScreen = async (req, res) => {
     const id = req.params.id
     try {
         await db_connect()
-        const deleted = await screenModel.findByIdAndDelete(id)
-        deleted ? res.status(204) : res.status(404).send("Screen not found")
+        const screen = await screenModel.findById(id)
+        if (screen) {
+            await seatModel.deleteMany({ screen: id }) //delete all seats in the screen
+            await screenModel.findByIdAndDelete(id)
+            res.status(204).send()
+        } else {
+            res.status(404).send("Screen not found")
+        }
     } catch (error) {
         res.status(500).send(error.message)
     } finally {
